refactor: split ConnectionFormView and MainContentView into extensions
ConnectionFormView.swift (1321 → 284 lines):
- +GeneralTab (218), +Footer (135), +Helpers (726)

MainContentView.swift (1137 → 453 lines):
- +EventHandlers (299), +Setup (234), +Helpers (129), +Modifiers (71)
- +Bindings (128) already existed

7 new extension files. No API changes.
