+++ TablePro/Views/Main/Extensions/MainContentView+EventHandlers.swift
@@ -1,0 +1,299 @@
+//
+//  MainContentView+EventHandlers.swift
+//  TablePro
+//
+//  Extension containing event handler methods for MainContentView.
+//  Extracted to reduce main view complexity.
+//
+
+import SwiftUI
+
+extension MainContentView {
+    // MARK: - Event Handlers
+
+    func handleTabSelectionChange(from oldTabId: UUID?, to newTabId: UUID?) {
+        coordinator.handleTabChange(
+            from: oldTabId,
+            to: newTabId,
+            selectedRowIndices: &selectedRowIndices,
+            tabs: tabManager.tabs
+        )
+
+        updateWindowTitleAndFileState()
+
+        // Sync sidebar selection to match the newly selected tab.
+        // Critical for new native windows: localSelectedTables starts empty,
+        // and this is the only place that can seed it from the restored tab.
+        syncSidebarToCurrentTab()
+
+        // Persist tab selection explicitly (skip during teardown)
+        guard !coordinator.isTearingDown else { return }
+        coordinator.persistence.saveNow(
+            tabs: tabManager.tabs,
+            selectedTabId: newTabId
+        )
+    }
+
+    func handleTabsChange(_ newTabs: [QueryTab]) {
+        updateWindowTitleAndFileState()
+
+        // Don't persist during teardown — SwiftUI may fire onChange with empty tabs
+        // as the view is being deallocated
+        guard !coordinator.isTearingDown else { return }
+        guard !coordinator.isUpdatingColumnLayout else { return }
+
+        // Promote preview tab if user has interacted with it
+        if let tab = tabManager.selectedTab, tab.isPreview, tab.hasUserInteraction {
+            coordinator.promotePreviewTab()
+        }
+
+        // Persist tab changes (exclude preview tabs from persistence)
+        let persistableTabs = newTabs.filter { !$0.isPreview }
+        if persistableTabs.isEmpty {
+            coordinator.persistence.clearSavedState()
+        } else {
+            let normalizedSelectedId =
+                persistableTabs.contains(where: { $0.id == tabManager.selectedTabId })
+                ? tabManager.selectedTabId : persistableTabs.first?.id
+            coordinator.persistence.saveNow(
+                tabs: persistableTabs,
+                selectedTabId: normalizedSelectedId
+            )
+        }
+    }
+
+    func handleColumnsChange(newColumns: [String]?) {
+        // Skip during tab switch — handleTabChange already configures the change manager
+        guard !coordinator.isHandlingTabSwitch else { return }
+
+        // Prune hidden columns that no longer exist in results
+        if let newColumns = newColumns {
+            coordinator.pruneHiddenColumns(currentColumns: newColumns)
+        }
+
+        guard let newColumns = newColumns, !newColumns.isEmpty,
+            let tab = tabManager.selectedTab,
+            !changeManager.hasChanges
+        else { return }
+
+        // Reconfigure if columns changed OR table name changed (switching tables)
+        let columnsChanged = changeManager.columns != newColumns
+        let tableChanged = changeManager.tableName != (tab.tableName ?? "")
+
+        guard columnsChanged || tableChanged else { return }
+
+        changeManager.configureForTable(
+            tableName: tab.tableName ?? "",
+            columns: newColumns,
+            primaryKeyColumn: newColumns.first,
+            databaseType: connection.type
+        )
+    }
+
+    func handleTableSelectionChange(
+        from oldTables: Set<TableInfo>, to newTables: Set<TableInfo>
+    ) {
+        let action = TableSelectionAction.resolve(oldTables: oldTables, newTables: newTables)
+
+        guard case .navigate(let tableName, let isView) = action else {
+            AppState.shared.hasTableSelection = !newTables.isEmpty
+            return
+        }
+
+        // Only navigate when this is the focused window.
+        // Prevents feedback loops when shared sidebar state syncs across native tabs.
+        guard isKeyWindow else {
+            AppState.shared.hasTableSelection = !newTables.isEmpty
+            return
+        }
+
+        let isPreviewMode = AppSettingsManager.shared.tabs.enablePreviewTabs
+        let hasPreview = WindowLifecycleMonitor.shared.previewWindow(for: connection.id) != nil
+
+        let result = SidebarNavigationResult.resolve(
+            clickedTableName: tableName,
+            currentTabTableName: tabManager.selectedTab?.tableName,
+            hasExistingTabs: !tabManager.tabs.isEmpty,
+            isPreviewTabMode: isPreviewMode,
+            hasPreviewTab: hasPreview
+        )
+
+        switch result {
+        case .skip:
+            AppState.shared.hasTableSelection = !newTables.isEmpty
+            return
+        case .openInPlace:
+            selectedRowIndices = []
+            coordinator.openTableTab(tableName, isView: isView)
+        case .revertAndOpenNewWindow:
+            coordinator.openTableTab(tableName, isView: isView)
+        case .replacePreviewTab, .openNewPreviewTab:
+            coordinator.openTableTab(tableName, isView: isView)
+        }
+
+        AppState.shared.hasTableSelection = !newTables.isEmpty
+    }
+
+    /// Keep sidebar selection in sync with the current window's tab.
+    /// Only writes when the value actually changes, preventing spurious onChange triggers.
+    /// Navigation safety is guaranteed by `SidebarNavigationResult.resolve` returning `.skip`
+    /// when the selected table matches the current tab.
+    func syncSidebarToCurrentTab() {
+        let target: Set<TableInfo>
+        if let currentTableName = tabManager.selectedTab?.tableName,
+            let match = tables.first(where: { $0.name == currentTableName })
+        {
+            target = [match]
+        } else {
+            target = []
+        }
+        if sidebarState.selectedTables != target {
+            // Don't clear sidebar selection while the table list is still loading.
+            // Clearing it prematurely triggers SidebarSyncAction to re-select on
+            // tables load, causing a double-navigation race condition.
+            if target.isEmpty && tables.isEmpty { return }
+            sidebarState.selectedTables = target
+        }
+    }
+
+    // MARK: - Sidebar Edit Handling
+
+    func updateSidebarEditState() {
+        guard let tab = coordinator.tabManager.selectedTab,
+            !selectedRowIndices.isEmpty
+        else {
+            rightPanelState.editState.fields = []
+            rightPanelState.editState.onFieldChanged = nil
+            return
+        }
+
+        var allRows: [[String?]] = []
+        for index in selectedRowIndices.sorted() {
+            if index < tab.resultRows.count {
+                allRows.append(tab.resultRows[index])
+            }
+        }
+
+        // Enrich column types with loaded enum values from Phase 2b
+        var columnTypes = tab.columnTypes
+        for (i, col) in tab.resultColumns.enumerated() where i < columnTypes.count {
+            if let values = tab.columnEnumValues[col], !values.isEmpty {
+                let ct = columnTypes[i]
+                if ct.isEnumType {
+                    columnTypes[i] = .enumType(rawType: ct.rawType, values: values)
+                } else if ct.isSetType {
+                    columnTypes[i] = .set(rawType: ct.rawType, values: values)
+                }
+            }
+        }
+
+        // Clear stale sidebar edits after refresh/discard
+        if !changeManager.hasChanges {
+            rightPanelState.editState.clearEdits()
+        }
+
+        // Collect columns modified in data grid so sidebar shows green dots
+        var modifiedColumns = Set<Int>()
+        for rowIndex in selectedRowIndices {
+            modifiedColumns.formUnion(changeManager.getModifiedColumnsForRow(rowIndex))
+        }
+
+        let excludedNames: Set<String>
+        if let tableName = tab.tableName {
+            excludedNames = Set(coordinator.columnExclusions(for: tableName).map(\.columnName))
+        } else {
+            excludedNames = []
+        }
+
+        rightPanelState.editState.configure(
+            selectedRowIndices: selectedRowIndices,
+            allRows: allRows,
+            columns: tab.resultColumns,
+            columnTypes: columnTypes,
+            externallyModifiedColumns: modifiedColumns,
+            excludedColumnNames: excludedNames
+        )
+
+        guard isSidebarEditable else {
+            rightPanelState.editState.onFieldChanged = nil
+            return
+        }
+
+        let capturedCoordinator = coordinator
+        let capturedEditState = rightPanelState.editState
+        rightPanelState.editState.onFieldChanged = { columnIndex, newValue in
+            guard let tab = capturedCoordinator.tabManager.selectedTab else { return }
+            let columnName =
+                columnIndex < tab.resultColumns.count ? tab.resultColumns[columnIndex] : ""
+
+            for rowIndex in capturedEditState.selectedRowIndices {
+                guard rowIndex < tab.resultRows.count else { continue }
+                let originalRow = tab.resultRows[rowIndex]
+
+                // Use full (lazy-loaded) original value if available, not truncated row data
+                let oldValue: String?
+                if columnIndex < capturedEditState.fields.count,
+                    !capturedEditState.fields[columnIndex].isTruncated
+                {
+                    oldValue = capturedEditState.fields[columnIndex].originalValue
+                } else {
+                    oldValue = columnIndex < originalRow.count ? originalRow[columnIndex] : nil
+                }
+
+                capturedCoordinator.changeManager.recordCellChange(
+                    rowIndex: rowIndex,
+                    columnIndex: columnIndex,
+                    columnName: columnName,
+                    oldValue: oldValue,
+                    newValue: newValue,
+                    originalRow: originalRow
+                )
+            }
+        }
+
+        // Lazy-load full values for excluded columns when a single row is selected
+        if !excludedNames.isEmpty,
+            selectedRowIndices.count == 1,
+            let tableName = tab.tableName,
+            let pkColumn = tab.primaryKeyColumn,
+            let rowIndex = selectedRowIndices.first,
+            rowIndex < tab.resultRows.count
+        {
+            let row = tab.resultRows[rowIndex]
+            if let pkColIndex = tab.resultColumns.firstIndex(of: pkColumn),
+                pkColIndex < row.count,
+                let pkValue = row[pkColIndex]
+            {
+                let excludedList = Array(excludedNames)
+
+                lazyLoadTask?.cancel()
+                lazyLoadTask = Task { @MainActor in
+                    let expectedRowIndex = rowIndex
+                    do {
+                        let fullValues =
+                            try await capturedCoordinator.fetchFullValuesForExcludedColumns(
+                                tableName: tableName,
+                                primaryKeyColumn: pkColumn,
+                                primaryKeyValue: pkValue,
+                                excludedColumnNames: excludedList
+                            )
+                        guard !Task.isCancelled,
+                            capturedEditState.selectedRowIndices.count == 1,
+                            capturedEditState.selectedRowIndices.first == expectedRowIndex
+                        else { return }
+                        capturedEditState.applyFullValues(fullValues)
+                    } catch {
+                        guard !Task.isCancelled,
+                            capturedEditState.selectedRowIndices.count == 1,
+                            capturedEditState.selectedRowIndices.first == expectedRowIndex
+                        else { return }
+                        for i in 0..<capturedEditState.fields.count
+                        where capturedEditState.fields[i].isLoadingFullValue {
+                            capturedEditState.fields[i].isLoadingFullValue = false
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
