+++ TablePro/Views/Main/MainContentView.swift
@@ -5,6 +5,13 @@
 //  Main content view combining query editor and results table.
 //  Refactored to use coordinator pattern for business logic separation.
 //
+//  Extensions:
+//  - MainContentView+Bindings.swift — computed bindings and trigger types
+//  - MainContentView+EventHandlers.swift — tab/table selection, sidebar edit handling
+//  - MainContentView+Setup.swift — initialization, command actions, database switching
+//  - MainContentView+Helpers.swift — helper methods, inspector context
+//  - MainContentView+Modifiers.swift — toolbar tint, focused command actions, preview
+//
 
 import Combine
 import SwiftUI
@@ -38,22 +45,22 @@ struct MainContentView: View {
     // MARK: - Local State
 
     @State var selectedRowIndices: Set<Int> = []
-    @State private var previousSelectedTabId: UUID?
-    @State private var previousSelectedTables: Set<TableInfo> = []
-    @State private var editingCell: CellPosition?
-    @State private var commandActions: MainContentCommandActions?
-    @State private var queryResultsSummaryCache: (tabId: UUID, version: Int, summary: String?)?
-    @State private var inspectorUpdateTask: Task<Void, Never>?
-    @State private var lazyLoadTask: Task<Void, Never>?
-    @State private var pendingTabSwitch: Task<Void, Never>?
-    @State private var evictionTask: Task<Void, Never>?
+    @State var previousSelectedTabId: UUID?
+    @State var previousSelectedTables: Set<TableInfo> = []
+    @State var editingCell: CellPosition?
+    @State var commandActions: MainContentCommandActions?
+    @State var queryResultsSummaryCache: (tabId: UUID, version: Int, summary: String?)?
+    @State var inspectorUpdateTask: Task<Void, Never>?
+    @State var lazyLoadTask: Task<Void, Never>?
+    @State var pendingTabSwitch: Task<Void, Never>?
+    @State var evictionTask: Task<Void, Never>?
     /// Stable identifier for this window in WindowLifecycleMonitor
-    @State private var windowId = UUID()
-    @State private var hasInitialized = false
+    @State var windowId = UUID()
+    @State var hasInitialized = false
     /// Tracks whether this view's window is the key (focused) window
-    @State private var isKeyWindow = false
+    @State var isKeyWindow = false
     /// Reference to this view's NSWindow for filtering notifications
-    @State private var viewWindow: NSWindow?
+    @State var viewWindow: NSWindow?
 
     /// Grace period for onDisappear: SwiftUI fires onDisappear transiently
     /// during tab group merges, then re-fires onAppear shortly after.
@@ -443,695 +450,4 @@ struct MainContentView: View {
             }
         )
     }
-
-    // MARK: - Initialization
-
-    private func initializeAndRestoreTabs() async {
-        guard !hasInitialized else { return }
-        hasInitialized = true
-        Task { await coordinator.loadSchemaIfNeeded() }
-
-        guard let payload else {
-            await handleRestoreOrDefault()
-            return
-        }
-
-        switch payload.intent {
-        case .openContent:
-            if payload.skipAutoExecute { return }
-            if let selectedTab = tabManager.selectedTab,
-                selectedTab.tabType == .table,
-                !selectedTab.query.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-            {
-                if let session = DatabaseManager.shared.activeSessions[connection.id],
-                    session.isConnected
-                {
-                    if !selectedTab.databaseName.isEmpty,
-                        selectedTab.databaseName != session.activeDatabase
-                    {
-                        Task { await coordinator.switchDatabase(to: selectedTab.databaseName) }
-                    } else {
-                        // columns is [] on initial load — buildFilteredQuery uses SELECT *
-                        if !selectedTab.filterState.appliedFilters.isEmpty,
-                            let tableName = selectedTab.tableName,
-                            let tabIndex = tabManager.selectedTabIndex
-                        {
-                            let filteredQuery = coordinator.queryBuilder.buildFilteredQuery(
-                                tableName: tableName,
-                                filters: selectedTab.filterState.appliedFilters,
-                                columns: [],
-                                limit: selectedTab.pagination.pageSize,
-                                offset: selectedTab.pagination.currentOffset
-                            )
-                            tabManager.tabs[tabIndex].query = filteredQuery
-                        }
-                        if let tableName = selectedTab.tableName {
-                            coordinator.restoreColumnLayoutForTable(tableName)
-                        }
-                        coordinator.executeTableTabQueryDirectly()
-                    }
-                } else {
-                    // Reactive path: fires via onChange(of: sessionVersion) when connection is ready
-                    coordinator.needsLazyLoad = true
-                }
-            }
-            if let sourceURL = payload.sourceFileURL {
-                WindowLifecycleMonitor.shared.registerSourceFile(sourceURL, windowId: windowId)
-            }
-
-        case .newEmptyTab:
-            return
-
-        case .restoreOrDefault:
-            await handleRestoreOrDefault()
-        }
-    }
-
-    private func handleRestoreOrDefault() async {
-        if WindowLifecycleMonitor.shared.hasOtherWindows(for: connection.id, excluding: windowId) {
-            if tabManager.tabs.isEmpty {
-                tabManager.addTab(databaseName: connection.database)
-            }
-            return
-        }
-
-        let result = await coordinator.persistence.restoreFromDisk()
-        if !result.tabs.isEmpty {
-            var restoredTabs = result.tabs
-            for i in restoredTabs.indices where restoredTabs[i].tabType == .table {
-                if let tableName = restoredTabs[i].tableName {
-                    restoredTabs[i].query = QueryTab.buildBaseTableQuery(
-                        tableName: tableName,
-                        databaseType: connection.type,
-                        schemaName: restoredTabs[i].schemaName
-                    )
-                }
-            }
-
-            let selectedId = result.selectedTabId
-            let selectedIndex = restoredTabs.firstIndex(where: { $0.id == selectedId }) ?? 0
-
-            let selectedTab = restoredTabs[selectedIndex]
-            tabManager.tabs = [selectedTab]
-            tabManager.selectedTabId = selectedTab.id
-
-            let remainingTabs = restoredTabs.enumerated()
-                .filter { $0.offset != selectedIndex }
-                .map(\.element)
-
-            if !remainingTabs.isEmpty {
-                Task { @MainActor in
-                    for tab in remainingTabs {
-                        let restorePayload = EditorTabPayload(
-                            from: tab, connectionId: connection.id, skipAutoExecute: true)
-                        WindowOpener.shared.openNativeTab(restorePayload)
-                    }
-                    viewWindow?.makeKeyAndOrderFront(nil)
-                }
-            }
-
-            if selectedTab.tabType == .table,
-                !selectedTab.query.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-            {
-                if let session = DatabaseManager.shared.activeSessions[connection.id],
-                    session.isConnected
-                {
-                    if !selectedTab.databaseName.isEmpty,
-                        selectedTab.databaseName != session.activeDatabase
-                    {
-                        Task { await coordinator.switchDatabase(to: selectedTab.databaseName) }
-                    } else {
-                        if let tableName = selectedTab.tableName {
-                            coordinator.restoreColumnLayoutForTable(tableName)
-                        }
-                        coordinator.executeTableTabQueryDirectly()
-                    }
-                } else {
-                    coordinator.needsLazyLoad = true
-                }
-            }
-        }
-    }
-
-    // MARK: - Command Actions Setup
-
-    private func updateToolbarPendingState() {
-        let hasDataChanges =
-            changeManager.hasChanges
-            || !pendingTruncates.isEmpty
-            || !pendingDeletes.isEmpty
-            || toolbarState.hasStructureChanges
-        let hasFileChanges = tabManager.selectedTab?.isFileDirty ?? false
-        toolbarState.hasDataPendingChanges = hasDataChanges
-        toolbarState.hasPendingChanges = hasDataChanges || hasFileChanges
-    }
-
-    /// Update window title, proxy icon, and dirty dot based on the selected tab.
-    private func updateWindowTitleAndFileState() {
-        let selectedTab = tabManager.selectedTab
-        if selectedTab?.tabType == .createTable {
-            windowTitle = String(localized: "Create Table")
-        } else if let fileURL = selectedTab?.sourceFileURL {
-            windowTitle = fileURL.deletingPathExtension().lastPathComponent
-        } else {
-            let langName = PluginManager.shared.queryLanguageName(for: connection.type)
-            let queryLabel = "\(langName) Query"
-            windowTitle = selectedTab?.tableName
-                ?? (tabManager.tabs.isEmpty ? connection.name : queryLabel)
-        }
-        viewWindow?.representedURL = selectedTab?.sourceFileURL
-        viewWindow?.isDocumentEdited = selectedTab?.isFileDirty ?? false
-    }
-
-    /// Configure the hosting NSWindow — called by WindowAccessor when the window is available.
-    private func configureWindow(_ window: NSWindow) {
-        let isPreview = tabManager.selectedTab?.isPreview ?? payload?.isPreview ?? false
-        if isPreview {
-            window.subtitle = "\(connection.name) — Preview"
-        } else {
-            window.subtitle = connection.name
-        }
-
-        let resolvedId = WindowOpener.tabbingIdentifier(for: connection.id)
-        window.tabbingIdentifier = resolvedId
-        window.tabbingMode = .preferred
-        coordinator.windowId = windowId
-
-        WindowLifecycleMonitor.shared.register(
-            window: window,
-            connectionId: connection.id,
-            windowId: windowId,
-            isPreview: isPreview
-        )
-        viewWindow = window
-        coordinator.contentWindow = window
-        isKeyWindow = window.isKeyWindow
-
-        if let payloadId = payload?.id {
-            WindowOpener.shared.acknowledgePayload(payloadId)
-        }
-
-        // Native proxy icon (Cmd+click shows path in Finder) and dirty dot
-        window.representedURL = tabManager.selectedTab?.sourceFileURL
-        window.isDocumentEdited = tabManager.selectedTab?.isFileDirty ?? false
-
-        // Update command actions window reference now that it's available
-        commandActions?.window = window
-    }
-
-    private func setupCommandActions() {
-        let actions = MainContentCommandActions(
-            coordinator: coordinator,
-            filterStateManager: filterStateManager,
-            connection: connection,
-            selectedRowIndices: $selectedRowIndices,
-            selectedTables: Binding(
-                get: { sidebarState.selectedTables },
-                set: { sidebarState.selectedTables = $0 }
-            ),
-            pendingTruncates: $pendingTruncates,
-            pendingDeletes: $pendingDeletes,
-            tableOperationOptions: $tableOperationOptions,
-            rightPanelState: rightPanelState,
-            editingCell: $editingCell
-        )
-        actions.window = viewWindow
-        commandActions = actions
-    }
-
-    // MARK: - Database Switcher
-
-    private func switchDatabase(to database: String) {
-        Task {
-            await coordinator.switchDatabase(to: database)
-        }
-    }
-
-    // MARK: - Event Handlers
-
-    private func handleTabSelectionChange(from oldTabId: UUID?, to newTabId: UUID?) {
-        coordinator.handleTabChange(
-            from: oldTabId,
-            to: newTabId,
-            selectedRowIndices: &selectedRowIndices,
-            tabs: tabManager.tabs
-        )
-
-        updateWindowTitleAndFileState()
-
-        // Sync sidebar selection to match the newly selected tab.
-        // Critical for new native windows: localSelectedTables starts empty,
-        // and this is the only place that can seed it from the restored tab.
-        syncSidebarToCurrentTab()
-
-        // Persist tab selection explicitly (skip during teardown)
-        guard !coordinator.isTearingDown else { return }
-        coordinator.persistence.saveNow(
-            tabs: tabManager.tabs,
-            selectedTabId: newTabId
-        )
-    }
-
-    private func handleTabsChange(_ newTabs: [QueryTab]) {
-        updateWindowTitleAndFileState()
-
-        // Don't persist during teardown — SwiftUI may fire onChange with empty tabs
-        // as the view is being deallocated
-        guard !coordinator.isTearingDown else { return }
-        guard !coordinator.isUpdatingColumnLayout else { return }
-
-        // Promote preview tab if user has interacted with it
-        if let tab = tabManager.selectedTab, tab.isPreview, tab.hasUserInteraction {
-            coordinator.promotePreviewTab()
-        }
-
-        // Persist tab changes (exclude preview tabs from persistence)
-        let persistableTabs = newTabs.filter { !$0.isPreview }
-        if persistableTabs.isEmpty {
-            coordinator.persistence.clearSavedState()
-        } else {
-            let normalizedSelectedId =
-                persistableTabs.contains(where: { $0.id == tabManager.selectedTabId })
-                ? tabManager.selectedTabId : persistableTabs.first?.id
-            coordinator.persistence.saveNow(
-                tabs: persistableTabs,
-                selectedTabId: normalizedSelectedId
-            )
-        }
-    }
-
-    private func handleColumnsChange(newColumns: [String]?) {
-        // Skip during tab switch — handleTabChange already configures the change manager
-        guard !coordinator.isHandlingTabSwitch else { return }
-
-        // Prune hidden columns that no longer exist in results
-        if let newColumns = newColumns {
-            coordinator.pruneHiddenColumns(currentColumns: newColumns)
-        }
-
-        guard let newColumns = newColumns, !newColumns.isEmpty,
-            let tab = tabManager.selectedTab,
-            !changeManager.hasChanges
-        else { return }
-
-        // Reconfigure if columns changed OR table name changed (switching tables)
-        let columnsChanged = changeManager.columns != newColumns
-        let tableChanged = changeManager.tableName != (tab.tableName ?? "")
-
-        guard columnsChanged || tableChanged else { return }
-
-        changeManager.configureForTable(
-            tableName: tab.tableName ?? "",
-            columns: newColumns,
-            primaryKeyColumn: newColumns.first,
-            databaseType: connection.type
-        )
-    }
-
-    private func handleTableSelectionChange(
-        from oldTables: Set<TableInfo>, to newTables: Set<TableInfo>
-    ) {
-        let action = TableSelectionAction.resolve(oldTables: oldTables, newTables: newTables)
-
-        guard case .navigate(let tableName, let isView) = action else {
-            AppState.shared.hasTableSelection = !newTables.isEmpty
-            return
-        }
-
-        // Only navigate when this is the focused window.
-        // Prevents feedback loops when shared sidebar state syncs across native tabs.
-        guard isKeyWindow else {
-            AppState.shared.hasTableSelection = !newTables.isEmpty
-            return
-        }
-
-        let isPreviewMode = AppSettingsManager.shared.tabs.enablePreviewTabs
-        let hasPreview = WindowLifecycleMonitor.shared.previewWindow(for: connection.id) != nil
-
-        let result = SidebarNavigationResult.resolve(
-            clickedTableName: tableName,
-            currentTabTableName: tabManager.selectedTab?.tableName,
-            hasExistingTabs: !tabManager.tabs.isEmpty,
-            isPreviewTabMode: isPreviewMode,
-            hasPreviewTab: hasPreview
-        )
-
-        switch result {
-        case .skip:
-            AppState.shared.hasTableSelection = !newTables.isEmpty
-            return
-        case .openInPlace:
-            selectedRowIndices = []
-            coordinator.openTableTab(tableName, isView: isView)
-        case .revertAndOpenNewWindow:
-            coordinator.openTableTab(tableName, isView: isView)
-        case .replacePreviewTab, .openNewPreviewTab:
-            coordinator.openTableTab(tableName, isView: isView)
-        }
-
-        AppState.shared.hasTableSelection = !newTables.isEmpty
-    }
-
-    /// Keep sidebar selection in sync with the current window's tab.
-    /// Only writes when the value actually changes, preventing spurious onChange triggers.
-    /// Navigation safety is guaranteed by `SidebarNavigationResult.resolve` returning `.skip`
-    /// when the selected table matches the current tab.
-    private func syncSidebarToCurrentTab() {
-        let target: Set<TableInfo>
-        if let currentTableName = tabManager.selectedTab?.tableName,
-            let match = tables.first(where: { $0.name == currentTableName })
-        {
-            target = [match]
-        } else {
-            target = []
-        }
-        if sidebarState.selectedTables != target {
-            // Don't clear sidebar selection while the table list is still loading.
-            // Clearing it prematurely triggers SidebarSyncAction to re-select on
-            // tables load, causing a double-navigation race condition.
-            if target.isEmpty && tables.isEmpty { return }
-            sidebarState.selectedTables = target
-        }
-    }
-
-    // MARK: - Helper Methods
-
-    private func loadTableMetadataIfNeeded() async {
-        guard let tableName = currentTab?.tableName,
-            coordinator.tableMetadata?.tableName != tableName
-        else { return }
-        await coordinator.loadTableMetadata(tableName: tableName)
-    }
-
-    private func handleConnectionStatusChange() {
-        let sessions = DatabaseManager.shared.activeSessions
-        guard let session = sessions[connection.id] else { return }
-        if session.isConnected && coordinator.needsLazyLoad {
-            let hasPendingEdits =
-                changeManager.hasChanges
-                || (tabManager.selectedTab?.pendingChanges.hasChanges ?? false)
-            guard !hasPendingEdits else { return }
-            coordinator.needsLazyLoad = false
-            if let selectedTab = tabManager.selectedTab,
-                !selectedTab.databaseName.isEmpty,
-                selectedTab.databaseName != session.activeDatabase
-            {
-                Task { await coordinator.switchDatabase(to: selectedTab.databaseName) }
-            } else if let selectedTab = tabManager.selectedTab,
-                let tabSchema = selectedTab.schemaName,
-                !tabSchema.isEmpty,
-                tabSchema != session.currentSchema
-            {
-                // Restore schema on the driver without clearing tabs (unlike switchSchema which resets UI)
-                Task {
-                    await coordinator.restoreSchemaAndRunQuery(tabSchema)
-                }
-            } else {
-                coordinator.runQuery()
-            }
-        }
-        let mappedState = mapSessionStatus(session.status)
-        if mappedState != toolbarState.connectionState {
-            toolbarState.connectionState = mappedState
-        }
-    }
-
-    private func mapSessionStatus(_ status: ConnectionStatus) -> ToolbarConnectionState {
-        switch status {
-        case .connected: return .connected
-        case .connecting: return .executing
-        case .disconnected: return .disconnected
-        case .error: return .error("")
-        }
-    }
-
-    // MARK: - Sidebar Edit Handling
-
-    private func updateSidebarEditState() {
-        guard let tab = coordinator.tabManager.selectedTab,
-            !selectedRowIndices.isEmpty
-        else {
-            rightPanelState.editState.fields = []
-            rightPanelState.editState.onFieldChanged = nil
-            return
-        }
-
-        var allRows: [[String?]] = []
-        for index in selectedRowIndices.sorted() {
-            if index < tab.resultRows.count {
-                allRows.append(tab.resultRows[index])
-            }
-        }
-
-        // Enrich column types with loaded enum values from Phase 2b
-        var columnTypes = tab.columnTypes
-        for (i, col) in tab.resultColumns.enumerated() where i < columnTypes.count {
-            if let values = tab.columnEnumValues[col], !values.isEmpty {
-                let ct = columnTypes[i]
-                if ct.isEnumType {
-                    columnTypes[i] = .enumType(rawType: ct.rawType, values: values)
-                } else if ct.isSetType {
-                    columnTypes[i] = .set(rawType: ct.rawType, values: values)
-                }
-            }
-        }
-
-        // Clear stale sidebar edits after refresh/discard
-        if !changeManager.hasChanges {
-            rightPanelState.editState.clearEdits()
-        }
-
-        // Collect columns modified in data grid so sidebar shows green dots
-        var modifiedColumns = Set<Int>()
-        for rowIndex in selectedRowIndices {
-            modifiedColumns.formUnion(changeManager.getModifiedColumnsForRow(rowIndex))
-        }
-
-        let excludedNames: Set<String>
-        if let tableName = tab.tableName {
-            excludedNames = Set(coordinator.columnExclusions(for: tableName).map(\.columnName))
-        } else {
-            excludedNames = []
-        }
-
-        rightPanelState.editState.configure(
-            selectedRowIndices: selectedRowIndices,
-            allRows: allRows,
-            columns: tab.resultColumns,
-            columnTypes: columnTypes,
-            externallyModifiedColumns: modifiedColumns,
-            excludedColumnNames: excludedNames
-        )
-
-        guard isSidebarEditable else {
-            rightPanelState.editState.onFieldChanged = nil
-            return
-        }
-
-        let capturedCoordinator = coordinator
-        let capturedEditState = rightPanelState.editState
-        rightPanelState.editState.onFieldChanged = { columnIndex, newValue in
-            guard let tab = capturedCoordinator.tabManager.selectedTab else { return }
-            let columnName =
-                columnIndex < tab.resultColumns.count ? tab.resultColumns[columnIndex] : ""
-
-            for rowIndex in capturedEditState.selectedRowIndices {
-                guard rowIndex < tab.resultRows.count else { continue }
-                let originalRow = tab.resultRows[rowIndex]
-
-                // Use full (lazy-loaded) original value if available, not truncated row data
-                let oldValue: String?
-                if columnIndex < capturedEditState.fields.count,
-                    !capturedEditState.fields[columnIndex].isTruncated
-                {
-                    oldValue = capturedEditState.fields[columnIndex].originalValue
-                } else {
-                    oldValue = columnIndex < originalRow.count ? originalRow[columnIndex] : nil
-                }
-
-                capturedCoordinator.changeManager.recordCellChange(
-                    rowIndex: rowIndex,
-                    columnIndex: columnIndex,
-                    columnName: columnName,
-                    oldValue: oldValue,
-                    newValue: newValue,
-                    originalRow: originalRow
-                )
-            }
-        }
-
-        // Lazy-load full values for excluded columns when a single row is selected
-        if !excludedNames.isEmpty,
-            selectedRowIndices.count == 1,
-            let tableName = tab.tableName,
-            let pkColumn = tab.primaryKeyColumn,
-            let rowIndex = selectedRowIndices.first,
-            rowIndex < tab.resultRows.count
-        {
-            let row = tab.resultRows[rowIndex]
-            if let pkColIndex = tab.resultColumns.firstIndex(of: pkColumn),
-                pkColIndex < row.count,
-                let pkValue = row[pkColIndex]
-            {
-                let excludedList = Array(excludedNames)
-
-                lazyLoadTask?.cancel()
-                lazyLoadTask = Task { @MainActor in
-                    let expectedRowIndex = rowIndex
-                    do {
-                        let fullValues =
-                            try await capturedCoordinator.fetchFullValuesForExcludedColumns(
-                                tableName: tableName,
-                                primaryKeyColumn: pkColumn,
-                                primaryKeyValue: pkValue,
-                                excludedColumnNames: excludedList
-                            )
-                        guard !Task.isCancelled,
-                            capturedEditState.selectedRowIndices.count == 1,
-                            capturedEditState.selectedRowIndices.first == expectedRowIndex
-                        else { return }
-                        capturedEditState.applyFullValues(fullValues)
-                    } catch {
-                        guard !Task.isCancelled,
-                            capturedEditState.selectedRowIndices.count == 1,
-                            capturedEditState.selectedRowIndices.first == expectedRowIndex
-                        else { return }
-                        for i in 0..<capturedEditState.fields.count
-                        where capturedEditState.fields[i].isLoadingFullValue {
-                            capturedEditState.fields[i].isLoadingFullValue = false
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // MARK: - Inspector Context
-
-    /// Coalesces multiple onChange-triggered updates into a single deferred call.
-    /// During tab switch, onChange handlers fire 3-4x — this ensures we only rebuild once,
-    /// and defers the work so SwiftUI can render the tab switch first.
-    private func scheduleInspectorUpdate() {
-        inspectorUpdateTask?.cancel()
-        inspectorUpdateTask = Task { @MainActor in
-            try? await Task.sleep(for: .milliseconds(100))
-            guard !Task.isCancelled else { return }
-            updateSidebarEditState()
-            updateInspectorContext()
-        }
-    }
-
-    private func updateInspectorContext() {
-        inspectorContext = InspectorContext(
-            tableName: currentTab?.tableName,
-            tableMetadata: coordinator.tableMetadata,
-            selectedRowData: selectedRowDataForSidebar,
-            isEditable: isSidebarEditable,
-            isRowDeleted: isSelectedRowDeleted,
-            currentQuery: coordinator.tabManager.selectedTab?.query,
-            queryResults: cachedQueryResultsSummary()
-        )
-    }
-
-    private func cachedQueryResultsSummary() -> String? {
-        guard let tab = currentTab else { return nil }
-        if let cache = queryResultsSummaryCache,
-            cache.tabId == tab.id, cache.version == tab.resultVersion
-        {
-            return cache.summary
-        }
-        let summary = buildQueryResultsSummary()
-        queryResultsSummaryCache = (tabId: tab.id, version: tab.resultVersion, summary: summary)
-        return summary
-    }
-
-    private func buildQueryResultsSummary() -> String? {
-        guard let tab = currentTab,
-            !tab.resultColumns.isEmpty,
-            !tab.resultRows.isEmpty
-        else { return nil }
-
-        let columns = tab.resultColumns
-        let rows = tab.resultRows
-        let maxRows = 10
-        let displayRows = Array(rows.prefix(maxRows))
-
-        var lines: [String] = []
-        lines.append(columns.joined(separator: " | "))
-
-        for row in displayRows {
-            let values = columns.indices.map { i in
-                i < row.count ? (row[i] ?? "NULL") : "NULL"
-            }
-            lines.append(values.joined(separator: " | "))
-        }
-
-        if rows.count > maxRows {
-            lines.append("(showing \(maxRows) of \(rows.count) rows)")
-        }
-
-        return lines.joined(separator: "\n")
-    }
-}
-
-// MARK: - Toolbar Tint Modifier
-
-/// Applies a subtle color tint to the window toolbar when a connection color is set.
-private struct ToolbarTintModifier: ViewModifier {
-    let connectionColor: ConnectionColor
-
-    @ViewBuilder
-    func body(content: Content) -> some View {
-        if connectionColor.isDefault {
-            content
-        } else {
-            content
-                .toolbarBackground(connectionColor.color.opacity(0.12), for: .windowToolbar)
-                .toolbarBackground(.visible, for: .windowToolbar)
-        }
-    }
-}
-
-// MARK: - Focused Command Actions Modifier
-
-/// Conditionally publishes `MainContentCommandActions` as a focused scene value.
-/// `focusedSceneValue` requires a non-optional value, so this modifier
-/// only applies it when the actions object has been created.
-private struct FocusedCommandActionsModifier: ViewModifier {
-    let actions: MainContentCommandActions?
-
-    func body(content: Content) -> some View {
-        if let actions {
-            content.focusedSceneValue(\.commandActions, actions)
-        } else {
-            content
-        }
-    }
-}
-
-// MARK: - Preview
-
-#Preview("With Connection") {
-    let state = SessionStateFactory.create(
-        connection: DatabaseConnection.preview,
-        payload: nil
-    )
-    MainContentView(
-        connection: DatabaseConnection.preview,
-        payload: nil,
-        windowTitle: .constant("SQL Query"),
-        tables: .constant([]),
-        sidebarState: SharedSidebarState(),
-        pendingTruncates: .constant([]),
-        pendingDeletes: .constant([]),
-        tableOperationOptions: .constant([:]),
-        inspectorContext: .constant(.empty),
-        rightPanelState: RightPanelState(),
-        tabManager: state.tabManager,
-        changeManager: state.changeManager,
-        filterStateManager: state.filterStateManager,
-        toolbarState: state.toolbarState,
-        coordinator: state.coordinator
-    )
-    .frame(width: 1_000, height: 600)
 }
